+++ CrowdStrikeFalcon/CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2026-06-23 - 1.25.17
+
+### Fixed
+
+- Use a default connector configuration inheritance
+
+## 2026-05-19 - 1.25.16
+
+### Fixed
+
+- Set EVENTS_LAG metric to zero when there are no events to forward
+
+## 2026-05-15 - 1.25.15
+
+### Fixed
+
+- Prevent stream offset from regressing in cache when a batch with a lower offset is written after a higher one
+
 ## 2026-04-27 - 1.25.14
 
 ### Fixed